// src/shared/api/generated/models/updateAssignmentDto.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * VIO Stats Service API
 * API для управления компаниями, рекрутерами и вакансиями
 * OpenAPI spec version: 1.0
 */

export interface UpdateAssignmentDto {
  /** Количество необходимых резюме */
  requiredResumes?: number;
  /** Количество отправленных резюме */
  sentResumes?: number;
  /** Количество принятых резюме */
  acceptedResumes?: number;
  /** Количество отклоненных резюме */
  rejectedResumes?: number;
}